Some breaking changes.
Remove optional event handler feature signature, but implement the
ability to pass undefined as event handler. In that case the event
will not be added. But the error will be thrown if the event is not
specified in attach object. That should make optional events become
easy to use both with the ability to detect unintentional properti omitting.
The second feature introduses "?=" operator used only inside tag with
attributes that do not expect any value (e.g. "disable"). For less boilerplate
code the $yes and $no tokens where added for static attribute setting.
Plus from now on the string attribute with a resulting value being an
empty string will be removed from the DOM element. Without that, the tag
may end up having a bunch of attributes being set, but with no value,
which is indeed confusing. I'm sorry for the breaking changes being made.